+++ coreblocks/interface/keys.py
@@ -6,11 +6,11 @@ from transactron.lib.dependencies import SimpleKey, UnifierKey, ListKey
 from transactron.lib.transformers import MethodProduct
 from transactron import Method, TModule
 from coreblocks.peripherals.bus_adapter import BusMasterInterface
+from coreblocks.func_blocks.csr.csr_protocol import RegisteredCSRProtocol
 from amaranth import Signal
 
 if TYPE_CHECKING:
     from coreblocks.priv.csr.csr_instances import CSRInstances  # noqa: F401
-    from coreblocks.func_blocks.csr.csr_unit import RegisteredCSRProtocol  # noqa: F401
 
 __all__ = [
     "CommonBusDataKey",
@@ -97,7 +97,7 @@ class CoreStateKey(SimpleKey[Method]):
 
 
 @dataclass(frozen=True)
-class CSRListKey(ListKey[tuple[int, "RegisteredCSRProtocol"]]):
+class CSRListKey(ListKey[tuple[int, RegisteredCSRProtocol]]):
     """DependencyManager key collecting CSR registers globally as a list.
     Requires tuple of architectural CSR number to register and the register itself."""
